error messages for team page done

+++ Web/action/EquipeAction.php
@@ -4,6 +4,7 @@
 	class EquipeAction extends CommonAction {
 		public $jobs = [];
 		public $members = [];
+		public $team = [];
 		public $error = "ok";
 
 		public function __construct() {
@@ -17,12 +18,9 @@
 			//fetch data from db for modifications
 			$this->jobs = UserDAO::getJobs();
 			$this->members = UserDAO::getTeamMembers();
-			if(isset($_POST["getbio"])){
-				$_SESSION["bio"] = UserDAO::getBio($_POST["getbio"]);
-				unset($_POST["getbio"]);
-			}
+			$this->team = UserDAO::getTeam();
 
-			if(isset($_POST["newfullname"]) && isset($_POST["newjob"]) && isset($_POST["newbio"]) && !empty($_FILES["newphoto"])){
+			if((isset($_POST["newfullname"]) && $_POST["newfullname"] != "") && (isset($_POST["newjob"]) && $_POST["newjob"] != "none") && (isset($_POST["newbio"])  && $_POST["newbio"] != "")&& $_FILES["newphoto"]["size"] != 0){
 				try{
 					$target_dir = "images/equipe/";
 					$target_file = $target_dir . basename($_FILES["newphoto"]["name"]);
@@ -34,39 +32,38 @@
 								if($imageFileType == "jpg" || $imageFileType == "png" || $imageFileType == "jpeg" || $imageFileType == "gif" ){
 									if(move_uploaded_file($_FILES["newphoto"]["tmp_name"], $target_file)){
 										UserDAO::newTeamMember($_POST["newfullname"], $_POST["newjob"], $_POST["newbio"], $target_file);
+										header("location:equipe.php");
+										exit;
 									}
 									else{
-										$this->error = "UPLOAD_ERROR";
+										$this->error = "UPLOAD_ERROR_NEW";
 									}
 								}
 								else{
-									$this->error = "FILE_TYPE_BAD";
+									$this->error = "FILE_TYPE_BAD_NEW";
 								}
 
 							}
 							else{
-								$this->error = "FILE_TOO_LARGE";
+								$this->error = "FILE_TOO_LARGE_NEW";
 							}
 						}
 						else{
-							$this->error = "FILE_EXISTS";
+							$this->error = "FILE_EXISTS_NEW";
 						}
 					}
 					else{
-						$this->error = "FILE_NOT_IMAGE";
+						$this->error = "FILE_NOT_IMAGE_NEW";
 					}
 				}
 				catch(Exception $e){
-					$this->error = "EXCEPTION_NEW_MEMBER";
+					$this->error = "ERROR_ADDING_MEMBER";
 				}
 
 				unset($_POST["newfullname"]);
 				unset($_POST["newjob"]);
 				unset($_POST["newbio"]);
 				unset($_FILES["newphoto"]);
-				header("location:equipe.php");
-				exit;
-
 			}
 
 
@@ -83,7 +80,7 @@
 						unset($_POST["editbio"]);
 					}
 					
-					if(!empty($_FILES["editphoto"])){
+					if($_FILES["editphoto"]["size"] != 0){
 
 						$target_dir = "images/equipe/";
 						$target_file = $target_dir . basename($_FILES["editphoto"]["name"]);
@@ -99,19 +96,19 @@
 										UserDAO::editMemberPhoto($_POST["editname"], $target_file);
 									}
 									else{
-										$this->error = "UPLOAD_ERROR";
+										$this->error = "UPLOAD_ERROR_EDIT";
 									}
 								}
 								else{
-									$this->error = "FILE_TYPE_BAD";
+									$this->error = "FILE_TYPE_BAD_EDIT";
 								}
 							}
 							else{
-								$this->error = "FILE_TOO_LARGE";
+								$this->error = "FILE_TOO_LARGE_EDIT";
 							}
 						}
 						else{
-							$this->error = "FILE_NOT_IMAGE";
+							$this->error = "FILE_NOT_IMAGE_EDIT";
 						}
 					}		
 					
@@ -132,17 +129,17 @@
 
 
 
-			if(isset($_POST["deleteteam"])){
+			if(isset($_POST["deleteteam"]) && $_POST["deleteteam"] != "none"){
 				try{
 					UserDAO::deleteTeamMember($_POST["deleteteam"]);
+					header("location:equipe.php");
+					exit;
 				}
 				catch(Exception $e){
 					$this->error = "EXCEPTION_DELETE";
 				}
 				
 				unset($_POST["deleteteam"]);
-				header("location:equipe.php");
-				exit;
 			}
 		}
 	}
